refactor v1_common
