+++ backend/internal/v1/v1_common/response.go
@@ -1,4 +1,4 @@
-package v1common
+package v1_common
 
 /*
 Use this for any json response that just needs a simple message field.
