+++ backend/internal/v1/v1_common/types.go
@@ -1,1 +1,1 @@
-package v1common
+package v1_common
